Update ClassicSpa package version to 6.0.740
Updated the `VpnHood.App.Client.csproj` file to change the version of the `VpnHood.AppLib.Assets.ClassicSpa` package from `6.0.738` to `6.0.740` in the `<ItemGroup>` that applies when the `$(SolutionDir)\..\.user\use-local-spa.txt` file does not exist. No changes were made to the configuration for using a local reference to the `ClassicSpa` library.